requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.13.0=pypi_0
aiofiles=23.2.1=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
attrs=25.3.0=pypi_0
bitsandbytes=0.45.5=pypi_0
bzip2=1.0.8=hda65f42_9
ca-certificates=2026.2.25=hbd8a1cb_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.1.8=pypi_0
cmake=4.2.3=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
deepspeed=0.12.6=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.6.1=pypi_0
einops-exts=0.0.4=pypi_0
exceptiongroup=1.3.0=pypi_0
fastapi=0.115.12=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.18.0=pypi_0
flash-attn=2.7.3=pypi_0
fonttools=4.58.0=pypi_0
fsspec=2025.5.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gradio=4.16.0=pypi_0
gradio-client=0.8.1=pypi_0
h11=0.14.0=pypi_0
hf-xet=1.3.2=pypi_0
hjson=3.1.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.0=pypi_0
huggingface-hub=0.36.2=pypi_0
icu=78.2=h33c6efd_0
idna=3.10=pypi_0
importlib-resources=6.5.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.1=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
kiwisolver=1.4.8=pypi_0
latex2mathml=3.78.0=pypi_0
ld_impl_linux-64=2.45.1=default_hbd61a6d_101
libexpat=2.7.4=hecca717_0
libffi=3.5.2=h3435931_0
libgcc=15.2.0=he0feb66_18
libgcc-ng=15.2.0=h69a702a_18
libgomp=15.2.0=he0feb66_18
liblzma=5.8.2=hb03c661_0
liblzma-devel=5.8.2=hb03c661_0
libnsl=2.0.1=hb9d3cd8_1
libsqlite=3.52.0=hf4e2dac_0
libstdcxx=15.2.0=h934c35e_18
libstdcxx-ng=15.2.0=hdf11a46_18
libuuid=2.41.3=h5347b49_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
llava=1.2.2.post1=pypi_0
markdown-it-py=3.0.0=pypi_0
markdown2=2.5.3=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.3=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
narwhals=1.40.0=pypi_0
ncurses=6.5=h2d0b736_3
networkx=3.4.2=pypi_0
ninja=1.13.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py=12.575.51=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.9.41=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.6.1=h35e630c_1
orjson=3.10.18=pypi_0
packaging=26.0=pyhcf101f3_0
pandas=2.2.3=pypi_0
peft=0.15.2=pypi_0
pillow=10.4.0=pypi_0
pip=26.0.1=pyh8b19718_0
platformdirs=4.3.8=pypi_0
protobuf=6.31.0=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pydantic=2.11.5=pypi_0
pydantic-core=2.33.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.1=pypi_0
pynvml=12.0.0=pypi_0
pyparsing=3.2.3=pypi_0
python=3.10.20=h3c07f61_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
python_abi=3.10=8_cp310
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.3=h853b02a_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=14.0.0=pypi_0
rpds-py=0.25.1=pypi_0
ruff=0.11.11=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.15.3=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=2.29.1=pypi_0
setproctitle=1.3.6=pypi_0
setuptools=82.0.1=pypi_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.52.0=h04a0ce9_0
starlette=0.46.2=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.14.0=pypi_0
threadpoolctl=3.6.0=pypi_0
timm=0.6.13=pypi_0
tk=8.6.13=noxft_h366c992_103
tokenizers=0.21.4=pypi_0
tomlkit=0.12.0=pypi_0
torch=2.2.2=pypi_0
torchaudio=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.47.0=pypi_0
triton=2.2.0=pypi_0
typer=0.15.4=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.1=pypi_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025.2=pypi_0
urllib3=2.4.0=pypi_0
uvicorn=0.34.2=pypi_0
wandb=0.19.11=pypi_0
wavedrom=2.0.3.post3=pypi_0
websockets=11.0.3=pypi_0
wheel=0.46.3=pyhd8ed1ab_0
xz=5.8.2=ha02ee65_0
xz-gpl-tools=5.8.2=ha02ee65_0
xz-tools=5.8.2=hb03c661_0
zlib=1.3.1=hb9d3cd8_2
zstd=1.5.7=hb78ec9c_6
